Combine to file together & Cache each item & Add doi & Add pdf download link in description & Styling the description
