+++ lib/v2/nber/router.js
@@ -1,4 +1,4 @@
 module.exports = (router) => {
-    router.get('/papers', require('./papers'));
-    router.get('/news', require('./news'));
+    router.get('/papers', require('.'));
+    router.get('/news', require('.'));
 };
