--- lib/v2/nber/template/author.art
@@ -1,5 +1,0 @@
-{{ if authors }}
-{{ each authors author }}
-    <a href={{ author.link }}>{{ author.name }}</a>
-{{ /each }}
-{{ /if }}
